transit_backend/requirements.txt
django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
Pillow>=10.1.0
django-filter>=23.5
django-redis>=5.4.0
redis>=6.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.0
# Testing tools
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
# Code quality tools
flake8>=6.1.0
black>=23.11.0
isort>=5.12.0 